feat: add classes
