+++ index.html
@@ -39,7 +39,7 @@
         </article>
       </section>
       <section class="user-input">
-        <div class="save-button ">
+        <div class="save-button">
           <button type="button" name="button">Save Palette</button>
         </div>
         <div class="random-button">
@@ -56,7 +56,8 @@
         </figure>
       </section>
     </main>
-    <script type="text/javascript" src="scripts.js"></script>
-    <script type="text/javascript" src="classes.js"></script>
+    <script type="text/javascript" src="./src/scripts.js"></script>
+    <script type="text/javascript" src="./src/Color.js"></script>
+    <script type="text/javascript" src="./src/Palette.js"></script>
   </body>
 </html>
