+++ src/Color.js
@@ -1,0 +1,8 @@
+class Color {
+    constructor() {
+      this.color = ``;
+      this.locked =false;
+    }
+}
+  
+//   module.exports = Color;
