+++ src/Palette.js
@@ -1,0 +1,10 @@
+// var Color = require("./Color")
+
+class Palette {
+    constructor() {
+      this.colors = [];
+      this.id = Date.now();
+    }
+}
+
+    //   module.exports = Palette
